requirements.txt
Django==1.11.12
djangorestframework==3.8.2
django-cors-headers==2.2.0
django-filter==1.1.0
kafka-python==1.4.2
